v8.17

+++ DB/SQLQuery1.sql
@@ -323,7 +323,7 @@ CREATE TABLE Sys_TotalCycleCost --经营概览:周期总生产成本
 (
 	id int identity (1,1) primary KEY,
 	Name varchar(50)  NULL,--物品
-	Cost INT , --成本
+	Cost FLOAT , --成本
 	TotalCost FLOAT,--总成本(值保持一致，读取其中一个)
 	PrType INT , --下拉框类型(一周:1、两周:2、三周:3、一个月:4、半年:5、一年:6)--忽略
 	AcctDate DATE,--数据日期
@@ -345,7 +345,7 @@ CREATE TABLE Sys_CostByDepartment --经营概览:自制成本按部门分类
 (
 	id int identity (1,1) primary KEY,
 	Name varchar(50)  NULL,--物品
-	Cost INT , --成本
+	Cost FLOAT  , --成本
 	PrType INT , --下拉框类型(一周:1、两周:2、三周:3、一个月:4、半年:5、一年:6)--忽略
 	AcctDate DATE,--数据日期
 	CreationTime DATETIME not null default getdate(),--创建时间
@@ -365,7 +365,7 @@ CREATE TABLE Sys_DeliveryCompletionRate --经营概览:交期达成率趋势
 (
 	id int identity (1,1) primary KEY,
 	Month varchar(50)  NULL,--月
-	DeliveryRate INT , --交付率
+	DeliveryRate FLOAT , --交付率
 	CreationTime DATETIME not null default getdate(),--创建时间
 	IsEffective int DEFAULT 1 -- 0 无效 1 有效 1显示
 )
